Address committee feedback (round 1): fix real bugs + document the contract (#464)
chat isAtBottom bindable sync after send; schema-form schema-change reset via
SchemaFormBody child extraction; rating/pin-input value no-write-back documented
(JSDoc + MINOR changeset); real behavioral tests replace source-scan ones.
627 affected tests pass; components:check clean.
