+++ packages/components/src/components/schema-form/schema-form.test.ts
@@ -810,6 +810,102 @@ describe('SchemaForm — composed-control regressions', () => {
   });
 });
 
+describe('SchemaForm — schema-change resets form state; value is seed-only', () => {
+  afterEach(() => cleanup());
+
+  test('changing schema resets formValue, errors, and rendered fields', async () => {
+    // Regression: {#key schema} in the OLD code only reconciled the DOM — it did
+    // NOT reset the $state variables (formValue, errors, rawDrafts, arrayKeys)
+    // because they lived in the component script scope, which persists across
+    // key-block remounts. The fix extracts those into a child component so the
+    // key-block genuinely recreates $state.
+
+    const schema1 = {
+      type: 'object',
+      properties: { name: { type: 'string', title: 'Name' } },
+      required: ['name'],
+    };
+    const schema2 = {
+      type: 'object',
+      properties: {
+        email: { type: 'string', title: 'Email' },
+        age: { type: 'integer', title: 'Age' },
+      },
+      required: ['email'],
+    };
+
+    const { rerender, container } = render(SchemaForm, {
+      props: {
+        schema: schema1 as never,
+        value: { name: 'Ada' },
+      },
+    });
+    await flush();
+
+    // Schema 1 renders a Name field seeded with 'Ada'.
+    const nameInput = screen.getByLabelText(/Name/);
+    expect(nameInput.value).toBe('Ada');
+
+    // Simulate user editing.
+    await fireEvent.input(nameInput, { target: { value: 'Grace' } });
+    expect(nameInput.value).toBe('Grace');
+
+    // Switch to schema2. The form should reset: Name field gone, Email + Age appear,
+    // and form state (formValue, errors) reset to the new schema's initial state.
+    await rerender({ schema: schema2 as never });
+    await flush();
+
+    // Old field is gone.
+    expect(container.querySelector('[aria-label*="Name"]')).toBeNull();
+
+    // New fields appear.
+    const emailInput = screen.getByLabelText(/Email/);
+    const ageInput = screen.getByRole('textbox', { name: /Age/ });
+    expect(emailInput).toBeTruthy();
+    expect(ageInput).toBeTruthy();
+
+    // formValue reset: new fields start empty (or schema-seeded), not carrying
+    // stale values from schema1.
+    expect(emailInput.value).toBe('');
+  });
+
+  test('changing value with same schema does NOT reset form state (seed-only contract)', async () => {
+    // The value prop is a seed: the consumer owns form state after mount.
+    // Changing value externally must NOT reset in-progress user edits — that
+    // would silently discard work-in-progress, which is worse than a stale binding.
+
+    const schema = {
+      type: 'object',
+      properties: { name: { type: 'string', title: 'Name' } },
+      required: ['name'],
+    };
+
+    const { rerender } = render(SchemaForm, {
+      props: {
+        schema: schema as never,
+        value: { name: 'Initial' },
+      },
+    });
+    await flush();
+
+    const nameInput = screen.getByLabelText(/Name/);
+    expect(nameInput.value).toBe('Initial');
+
+    // User edits the field.
+    await fireEvent.input(nameInput, { target: { value: 'User edit' } });
+    expect(nameInput.value).toBe('User edit');
+
+    // Consumer changes the value prop externally (same schema) — e.g. the parent
+    // refetches data. The form should NOT reset to the new value because the user
+    // has in-progress edits.
+    await rerender({ schema: schema as never, value: { name: 'Server update' } });
+    await flush();
+
+    // User edit is preserved — value prop change did not reset formValue.
+    expect(nameInput.value).toBe('User edit');
+  });
+});
+
 describe('SchemaForm — initialization without write-back $effect', () => {
   afterEach(() => cleanup());
